els-ASAP auto commit file/tag ImageMagick-6.7.2.7-2.el6.src.rpm

+++ src/coders/info.c
@@ -17,7 +17,7 @@
 %                                 July 1992                                   %
 %                                                                             %
 %                                                                             %
-%  Copyright 1999-2009 ImageMagick Studio LLC, a non-profit organization      %
+%  Copyright 1999-2011 ImageMagick Studio LLC, a non-profit organization      %
 %  dedicated to making software imaging solutions freely available.           %
 %                                                                             %
 %  You may not use this file except in compliance with the License.  You may  %
@@ -87,10 +87,10 @@ static MagickBooleanType
 %
 %  The format of the RegisterINFOImage method is:
 %
-%      unsigned long RegisterINFOImage(void)
+%      size_t RegisterINFOImage(void)
 %
 */
-ModuleExport unsigned long RegisterINFOImage(void)
+ModuleExport size_t RegisterINFOImage(void)
 {
   MagickInfo
     *entry;
